marketplace: create new web application for HIFIS Cloud Cluster
Motivation:

According to the official website of HIFIS, the aims of the project
is to ensure an excellent information environment for outstanding
research in all Helmholtz research fields. To provide a seamless and
performance oriented IT-infrastructure in connecting knowledge from
all centres.

To achieve this objective, a user friendly interface will be required.
This interface will be an entry point for scientist and partners alike
to request and use various services provided by the Helmholtz research
centres.

Modification:

Add neccessary files.

Result:

An animated welcome page with a simple unit test is created.

Target: master