ci-test: add SonarQube integration to the project
**Motivation:**

* automatically execute checks with SonarQube after each commit in CI

**Modification:**

* modified package.json to execute checks and added the necessary configuration

**Result:**

* SonarQubeChecks are executed automatically

**Additional Remarks:**

This is a new PR after cleaning up my fork and adressing the changes requested in PR #2 (https://github.com/helmholtz-marketplace/helmholtz-marketplace-webapp/pull/2). In detail, I changed the following:

* moved SonarQube access token to a environment variable which is created from a Github secret (see `README.md` for this)
* limited changes in `package.json` to the single added line `"sonarqube-scanner": "^2.6.0"`

I did not change the script in `package.json` starting with `"posttest"` this is valid according to the docs and I tried using `"test:post` which is not executed. 

The comment on the first line in `ci-analysis.js` is necessary, because otherwise eslint complains about extraneous dependencies.