rename project files and contents
Motivation:

Initially, this project is named `HIFIS` marketplace however
HIFIS is just one of the incubators belonging to the Helmholtz
associate. Hence, we replace the hifis name with helmholtz; to
reflect the project core objective and the association it belong.

Modification:

Change all neccesary file names and contents that contain the
acronym `hifis` to one of the following:
- helmholtz
- helmholtz-marketplace-app
- helmholtz-marketplace-webapp

Result:

- No visible changes to the end-user
- All project files and contents now contain appropriate
project name