add build and release process
Motivation:

Since this project will be a dependency for other projects we are
working on, we need a well managed artifact creation. Also, the
deployment and releases of this artifact is necessity for this project.
We will be using Apache maven which is a well established and
robust software project management tool.

Modification:

- create a custom assembly descriptor called `hmw.xml` which is used
    by maven-assembly-plugin for easy build distribution.
- add a javascript module bundler called `rollup` to optimise code for
    production and minifies JS, html and lit-html templates.
- add support for build, release and deploy processes in the `pom.xml`
    file.

Result:

- a new artifact can be created for a new version
- easily deploy artifact to the dcache nexus repositories

Pull-request: #7